// viewer/api/api-hooks.ts
const baseUrl = process.env.NEXT_PUBLIC_DIAGNOSTIC_SERVER_API
  ? process.env.NEXT_PUBLIC_DIAGNOSTIC_SERVER_API
  : process.env.DIAGNOSTIC_SERVER_API
    ? process.env.NEXT_DIAGNOSTIC_SERVER_API
    : 'http://localhost:8080/v1/'

const client = createClient<paths>({
  baseUrl: baseUrl,
})
// ...
export type ObjectMetadata = components['schemas']['ObjectMetadata']
// ...
export type DopUnionTypeWithLayerInfo = {
  dop?: DopUnionType
  origin_layer_perma_id?: string
  origin_layer_ephemeral_id?: number
  origin_layer_short_name?: string
  origin_layer_type?: string
}
// ...
export type DataObjectProperty = components['schemas']['DataObjectProperty']
export type DtcDop = components['schemas']['DtcDop']
export type ComplexDop = components['schemas']['ComplexDop']
export type Multiplexer = components['schemas']['Multiplexer']
export type BasicStructure = components['schemas']['BasicStructure']
export type EnvironmentDataDescription =
  components['schemas']['EnvironmentDataDescription']
export type Field = components['schemas']['Field']
// DOP type definitions - Specific
export type EnvironmentData = components['schemas']['EnvironmentData']
export type Structure = components['schemas']['Structure']
export type StaticField = components['schemas']['StaticField']
export type DynamicEndmarkerField =
  components['schemas']['DynamicEndmarkerField']
export type DynamicLengthField = components['schemas']['DynamicLengthField']
export type EndOfPduField = components['schemas']['EndOfPduField']
// ...
export type UnitResolved = components['schemas']['UnitResolved']
export type PhysicalDimension = components['schemas']['PhysicalDimension']
export type CompuInternalToPhys = components['schemas']['CompuInternalToPhys']
export type CompuPhysToInternal = components['schemas']['CompuPhysToInternal']
export type Encoding = components['schemas']['Encoding']

export type DopUnionType =
  | DataObjectProperty
  | DtcDop
  | Multiplexer
  | EnvironmentDataDescription
  | EnvironmentData
  | Structure
  | StaticField
  | DynamicEndmarkerField
  | DynamicLengthField
  | EndOfPduField